Renaming TreeShinyHunt and adding latest changes to LeapGrinder
