+++ SerialPrograms/Source/PokemonLA/Panels_PokemonLA.cpp
@@ -28,8 +28,8 @@
 #include "Programs/ShinyHunting/PokemonLA_GalladeFinder.h"
 #include "Programs/ShinyHunting/PokemonLA_CrobatFinder.h"
 #include "Programs/ShinyHunting/PokemonLA_FroslassFinder.h"
+#include "Programs/ShinyHunting/PokemonLA_BurmyFinder.h"
 #include "Programs/ShinyHunting/PokemonLA_UnownFinder.h"
-#include "Programs/ShinyHunting/PokemonLA_TreeShinyHunting.h"
 #include "Programs/ShinyHunting/PokemonLA_ShinyHunt-FlagPin.h"
 #include "Programs/ShinyHunting/PokemonLA_PostMMOSpawnReset.h"
 #include "Programs/ShinyHunting/PokemonLA_ShinyHunt-CustomPath.h"
@@ -75,8 +75,8 @@ Panels::Panels(QTabWidget& parent, PanelListener& listener)
     add_program<CrobatFinder_Descriptor, CrobatFinder>();
     add_program<GalladeFinder_Descriptor, GalladeFinder>();
     add_program<FroslassFinder_Descriptor, FroslassFinder>();
+    add_program<BurmyFinder_Descriptor, BurmyFinder>();
     add_program<UnownFinder_Descriptor, UnownFinder>();
-    add_program<TreeShinyHunting_Descriptor, TreeShinyHunting>();
     add_program<ShinyHuntFlagPin_Descriptor, ShinyHuntFlagPin>();
     add_program<PostMMOSpawnReset_Descriptor, PostMMOSpawnReset>();
     if (GlobalSettings::instance().DEVELOPER_MODE){
